+++ internal/infrastructure/db/postgres/sqlc/query.sql
@@ -437,7 +437,7 @@ VALUES (@asset_id, @txid, @vout, @amount);
 SELECT * FROM asset WHERE asset.id = ANY($1::varchar[]);
 
 -- name: SelectAssetSupply :one
-SELECT (COALESCE(SUM(ap.amount), 0)::NUMERIC)::TEXT AS supply
+SELECT (COALESCE(SUM(ap.amount), 0))::TEXT AS supply
 FROM asset_projection ap
 INNER JOIN vtxo v ON v.txid = ap.txid AND v.vout = ap.vout
 WHERE ap.asset_id = $1 AND v.spent = false;
